workaround for broken `/bus/objects/:key` renterd API method
Recent versions of renterd broke the `/bus/objects/:key` method when using it to read a directory. As a temporary workaround we now use the `/bus/objects/list` instead and filter out entries not belonging to the requested directory.

This is costly, but for now the only way that works properly. Once https://github.com/SiaFoundation/renterd/issues/1492 has been resolved, we can switch to the new `/listobjects/` method.